+++ lib/Duration.zig
@@ -12,14 +12,14 @@ const maxDigits: usize = 99;
 
 /// Any duration represented in seconds.
 /// Do not modify directly.
-__sec: i64 = 0,
+__sec: i128 = 0,
 
-/// Fractional seconds in nanoseconds, always positive.
+/// Fractional seconds in nanoseconds, always >= 0.
 /// Do not modify directly.
-__nsec: u32 = 0, // fraction is always positive
+__nsec: u32 = 0,
 
 /// Create a duration from multiple of specific a timespan
-pub fn fromTimespanMultiple(n: i128, timespan: Timespan) Duration {
+pub fn fromTimespanMultiple(n: i64, timespan: Timespan) Duration {
     const ns: i128 = @as(i128, @intCast(@intFromEnum(timespan))) * n;
     return .{
         .__sec = @intCast(@divFloor(ns, 1_000_000_000)),
@@ -29,8 +29,10 @@ pub fn fromTimespanMultiple(n: i128, timespan: Timespan) Duration {
 
 /// Create a duration from an 'ISO8601 duration' style string, e.g.
 ///
-/// PT9H - 9 hours
-/// PT1.234S - 1 second, 234 milliseconds
+/// ```
+/// "PT9H" : 9 hours
+/// "PT1.234S" : 1 second, 234 milliseconds
+/// ```
 ///
 /// Restrictions:
 /// - Since the Duration type represents an absolute difference in time,
@@ -52,15 +54,33 @@ pub fn toTimespanMultiple(duration: Duration, timespan: Timespan) i128 {
     return @intCast(result);
 }
 
-/// Representation with fractional seconds
+/// Representation as seconds with fractional seconds
 pub fn totalSeconds(duration: Duration) f64 {
     const ns = duration.asNanoseconds();
     return @floatCast(@as(f128, @floatFromInt(ns)) / 1_000_000_000);
 }
 
+/// Representation as minutes with fractional minutes
+pub fn totalMinutes(duration: Duration) f64 {
+    const ns = duration.asNanoseconds();
+    return @floatCast(@as(f128, @floatFromInt(ns)) / 60_000_000_000);
+}
+
+/// Representation as hours with fractional hours
+pub fn totalHours(duration: Duration) f64 {
+    const ns = duration.asNanoseconds();
+    return @floatCast(@as(f128, @floatFromInt(ns)) / 3_600_000_000_000);
+}
+
+/// Representation as days with fractional days
+pub fn totalDays(duration: Duration) f64 {
+    const ns = duration.asNanoseconds();
+    return @floatCast(@as(f128, @floatFromInt(ns)) / 86_400_000_000_000);
+}
+
 /// Add a duration to another. Makes a new Duration.P
 pub fn add(this: Duration, other: Duration) Duration {
-    const s: i64 = this.__sec + other.__sec;
+    const s: i128 = this.__sec + other.__sec;
     const ns: u32 = this.__nsec + other.__nsec;
     return .{
         .__sec = s + @divFloor(ns, 1_000_000_000),
@@ -70,7 +90,7 @@ pub fn add(this: Duration, other: Duration) Duration {
 
 /// Subtract a duration from another. Makes a new Duration.
 pub fn sub(this: Duration, other: Duration) Duration {
-    var s: i64 = this.__sec - other.__sec;
+    var s: i128 = this.__sec - other.__sec;
     var ns: i32 = @as(i32, @intCast(this.__nsec)) - @as(i32, @intCast(other.__nsec));
     if (ns < 0) {
         s -= 1;
@@ -79,8 +99,8 @@ pub fn sub(this: Duration, other: Duration) Duration {
     return .{ .__sec = s, .__nsec = @intCast(ns) };
 }
 
-/// Convert a duration to seconds, don't forget the nanos.
-pub fn asSeconds(duration: Duration) i64 {
+/// Convert a duration to seconds, don't forget the nanos just because they're small 😎
+pub fn asSeconds(duration: Duration) i128 {
     if (duration.__nsec > 500_000_000) return duration.__sec + 1;
     return duration.__sec;
 }
@@ -90,7 +110,7 @@ pub fn asNanoseconds(duration: Duration) i128 {
     return duration.__sec * 1_000_000_000 + duration.__nsec;
 }
 
-/// Formatted printing for Duration type. Defaults to 'ISO8601 duration'-like format.
+/// Formatted printing for Duration type. Defaults to 'ISO8601-duration'-like format.
 pub fn format(
     duration: Duration,
     comptime fmt: []const u8,
@@ -121,8 +141,7 @@ pub fn format(
     if (rd.seconds == 0 and rd.nanoseconds == 0) return;
 
     var frac = duration.__nsec;
-    // truncate zeros from fractional part
-    if (frac > 0) {
+    if (frac > 0) { // truncate zeros from fractional part
         while (frac % 10 == 0) : (frac /= 10) {}
     }
 
@@ -207,7 +226,7 @@ pub const RelativeDelta = struct {
             .negative = if (duration.__sec < 0) true else false,
         };
 
-        var secs: u64 = @abs(duration.__sec);
+        var secs: u128 = @abs(duration.__sec);
         result.weeks = @truncate(secs / (86400 * 7));
         secs -|= @as(u64, result.weeks) * (86400 * 7);
 
@@ -225,7 +244,7 @@ pub const RelativeDelta = struct {
         return result;
     }
 
-    /// Make a Duration from a RelativeDelta
+    /// Make a Duration from a RelativeDelta.
     pub fn toDuration(reldelta: RelativeDelta) FormatError!Duration {
         if (reldelta.years != 0 or reldelta.months != 0) return FormatError.InvalidFormat;
         const total_secs: i64 = @as(i64, reldelta.weeks) * 7 * 86400 + //
@@ -279,6 +298,7 @@ pub const RelativeDelta = struct {
 
         // need flags to keep track of what has been parsed already,
         // and in which order.
+        //
         // quantity/token:  Y m W d T H M S
         // bit/order:       7 6 5 4 3 2 1 0
         var flags: u8 = 0;
@@ -306,7 +326,8 @@ pub const RelativeDelta = struct {
                     } else {
                         if (flags > 0b111111) return FormatError.InvalidFormat;
                         // if no 'T' was parsed before, flags 0, 1 and 2 must be 0:
-                        if (flags & 0b1000 == 0 and flags & 0b111 != 0) return FormatError.InvalidFormat;
+                        if (flags & 0b1000 == 0 and flags & 0b111 != 0)
+                            return FormatError.InvalidFormat;
                         idx -= 1;
                         flags |= 1 << 6;
                         const quantity = try parseAndAdvanceYmWdHM(u32, string, &idx);
@@ -328,7 +349,8 @@ pub const RelativeDelta = struct {
                 },
                 'D' => {
                     if (flags > 0b1111) return FormatError.InvalidFormat;
-                    if (flags & 0b1000 == 0 and flags & 0b111 != 0) return FormatError.InvalidFormat;
+                    if (flags & 0b1000 == 0 and flags & 0b111 != 0)
+                        return FormatError.InvalidFormat;
                     idx -= 1;
                     flags |= 1 << 4;
                     const quantity = try parseAndAdvanceYmWdHM(u32, string, &idx);
@@ -336,7 +358,8 @@ pub const RelativeDelta = struct {
                 },
                 'W' => {
                     if (flags > 0b11111) return FormatError.InvalidFormat;
-                    if (flags & 0b1000 == 0 and flags & 0b111 != 0) return FormatError.InvalidFormat;
+                    if (flags & 0b1000 == 0 and flags & 0b111 != 0)
+                        return FormatError.InvalidFormat;
                     idx -= 1;
                     flags |= 1 << 5;
                     const quantity = try parseAndAdvanceYmWdHM(u32, string, &idx);
@@ -344,7 +367,8 @@ pub const RelativeDelta = struct {
                 },
                 'Y' => {
                     if (flags & 1 << 7 != 0) return FormatError.InvalidFormat;
-                    if (flags & 0b1000 == 0 and flags & 0b111 != 0) return FormatError.InvalidFormat;
+                    if (flags & 0b1000 == 0 and flags & 0b111 != 0)
+                        return FormatError.InvalidFormat;
                     idx -= 1;
                     flags |= 1 << 7;
                     const quantity = try parseAndAdvanceYmWdHM(u32, string, &idx);
@@ -360,7 +384,7 @@ pub const RelativeDelta = struct {
 
 /// Backwards-looking parse chars from 'string' seconds and nanoseconds (sum),
 /// end index is the value of 'idx_ptr' when the function is called.
-/// start index is determined automatically.
+/// Start index is determined automatically.
 ///
 /// This is a procedure; it modifies input pointer 'sec' and 'nsec' values in-place.
 /// This way, we can work around the fact that there are no multiple-return functions
@@ -399,7 +423,7 @@ fn parseAndAdvanceS(string: []const u8, idx_ptr: *usize, sec: *u32, nsec: *u32) 
     }
 }
 
-/// backwards-looking parse chars from 'string' to int (base 10),
+/// Backwards-looking parse chars from 'string' to int (base 10),
 /// end index is the value of 'idx_ptr' when the function is called.
 /// start index is determined automatically.
 fn parseAndAdvanceYmWdHM(comptime T: type, string: []const u8, idx_ptr: *usize) FormatError!T {