+++ lib/UTCoffset.zig
@@ -13,21 +13,21 @@ const UTCoffset = @This();
 
 const cap_designation_data: usize = 6;
 
-/// offset from UTC should be in range -25h to +26h as specified by
+/// Offset from UTC should be in range `(-25h ... +26h)` as specified by
 /// RFC9636, sect. 3.2, TZif data block.
 pub const offset_range = [2]i32{ -89999, 93599 };
 
-/// seconds East of Greenwich
+/// Seconds East of Greenwich.
 seconds_east: i32 = 0,
 
 /// DST indicator - can only be determined if the UTC offset is derived
-/// from a tz rule for given datetime.
+/// from a tz rule (POSIX, TZif) for given datetime.
 is_dst: bool = false,
 
-// 'internal' data for the offset designation
+// 'internal'; data for the offset designation.
 __designation_data: [6:0]u8 = [6:0]u8{ 0, 0, 0, 0, 0, 0 },
 
-// TZif transitions index; < 0 means invalid
+// 'internal'; TZif transitions index; < 0 means invalid.
 __transition_index: i32 = -1,
 
 /// UTC is constant. Presumably.
@@ -38,8 +38,8 @@ pub const UTC = UTCoffset{
 /// Designation / abbreviated time zone name such as "CET" for
 /// Central European Time in Europe/Berlin, winter.
 ///
-/// Note that multiple time zones can share the same abbreviated name and are
-/// therefore ambiguous.
+/// Note that multiple time zones can share the same abbreviated name;
+/// the abbreviations are therefore ambiguous.
 pub fn designation(offset: *const UTCoffset) []const u8 {
     return std.mem.sliceTo(&offset.__designation_data, 0);
 }
@@ -68,8 +68,8 @@ pub fn fromSeconds(offset_sec_East: i32, name: []const u8, is_dst: bool) TzError
 pub fn atUnixtime(tz: *const Timezone, unixtime: i64) TzError!UTCoffset {
     switch (tz.rules) {
         .tzif => {
-            // if the tz only has one timetype (offset spec.), use this,
-            // otherwise try to determine it from the defined transitions
+            // If the tz only has one timetype (offset spec.): use this.
+            // Otherwise try to determine it from the defined transitions.
             const idx = if (tz.rules.tzif.timetypes.len == 1) -1 else findTransition(tz.rules.tzif.transitions, unixtime);
 
             const timet = switch (idx) {
@@ -100,9 +100,7 @@ pub fn atUnixtime(tz: *const Timezone, unixtime: i64) TzError!UTCoffset {
                 .__transition_index = idx,
             };
         },
-        .tzif_fixedsize => {
-            // if the tz only has one timetype (offset spec.), use this,
-            // otherwise try to determine it from the defined transitions
+        .tzif_fixedsize => { // same as .tzif but for the fixed-size structure
             const idx = if (tz.rules.tzif_fixedsize.timetypes.len == 1) -1 else findTransition(tz.rules.tzif_fixedsize.transitions, unixtime);
 
             const timet = switch (idx) {
@@ -111,17 +109,12 @@ pub fn atUnixtime(tz: *const Timezone, unixtime: i64) TzError!UTCoffset {
                     break :blk tz.rules.tzif_fixedsize.timetypes[0];
                 },
 
-                // Unix time exceeds defined range of transitions => use POSIX from tzif footer
                 -2 => blk: {
-                    // check the POSIX TZ from the footer.
                     const psxtz = psx.parsePosixTzString(tz.rules.tzif_fixedsize.footer.?) catch return TzError.InvalidPosixTz;
-                    // If it has DST, make a UTC offset directly
                     if (psxtz.dst_offset) |_| return psxtz.utcOffsetAt(unixtime);
-                    // ...otherwise use existing timetype
                     break :blk tz.rules.tzif_fixedsize.transitions[tz.rules.tzif_fixedsize.transitions.len - 1].timetype.*;
                 },
 
-                // Unix time precedes defined range of transitions => use first entry in timetypes (likely a LMT)
                 -3 => tz.rules.tzif_fixedsize.timetypes[0],
                 else => tz.rules.tzif_fixedsize.transitions[@intCast(idx)].timetype.*,
             };
@@ -138,6 +131,9 @@ pub fn atUnixtime(tz: *const Timezone, unixtime: i64) TzError!UTCoffset {
     }
 }
 
+/// Custom formatter for the Offset struct.
+/// `fmt` is ignored.
+/// `options` can be used to modify precision.
 pub fn format(
     offset: UTCoffset,
     comptime fmt: []const u8,
@@ -167,10 +163,13 @@ pub fn format(
 }
 
 /// Get the index of the UTC offset transition equal to or less than the given target.
+///
 /// Invalid value indicators:
+/// ```
 /// -1 : transition array has no elements (single offset tz)
 /// -2 : target is larger than last transition element
 /// -3 : target is smaller than first transition element
+/// ```
 fn findTransition(array: []const tzif.Transition, target: i64) i32 {
     if (array.len == 0) return -1;
     // we know that transitions in 'array' are sorted, so we can check first and last indices.